Quantity from Double to Int

+++ src/longpipe/AddPipe.java
@@ -550,7 +550,7 @@ public class AddPipe extends javax.swing.JDialog {
             tfErrorOutput.setForeground(darkGreen);
             tfErrorOutput.setText("This pipe is stocked and is of type " + type);
             PipeMain pipePrice = createPipe();
-            tfTotalCostOutput.setText(String.format("%.2f", pipePrice.getPrice() * Double.parseDouble(tfQuantityInput.getText())));
+            tfTotalCostOutput.setText(String.format("%.2f", pipePrice.getPrice() * Integer.parseInt(tfQuantityInput.getText())));
             tfCostPerPipeOutput.setText(String.format("%.2f", pipePrice.getPrice()));
         }
     }//GEN-LAST:event_btnTestValidActionPerformed
